refactor: move graphqlclient from useApi to context
M2-354

this is so you can query graphql endpoints in plugins

refactor: useApi to use client from context

refactor: move dynamic headers to client instance

I don't want to keep the append the headers in useApi because if someone
wants to pull the graphql clients from the context - e.g. when
requesting something in a plugin - they'll have to reimplement the
cookie logic again

this commit is a checkpoint commit - the '' fallback in graphqlClient
isn't ideal because the header shouldn't be appended to the request if
it's empty. Getters are limited int this way. I'll try something with Proxy

refactor: proxy workaround for dynamic headers

refactor: use proxies as workaround for dynamic headers

refactor: add template types to useApi

refactor: remove references to customApolloHttpLinkOptions

fix: fix spread

refactor: code review changes