chore: eslint composables and api-client properly
There are now 2 tsconfigs: one (.eslint.json) that tells eslint which
files to scan, and tsconfig.json which is related to building the library

If you try to lint a file that's not part of any TS project, ESLint will
complain that it needs to be part of some TS project - it can't be
"exclude"d etc.

But at the same time, I don't want to mess with the types distributed by
the api-client package (ie. I want to lin .spec.ts files but I don't
want them to appear in the library's dist). Hence the
.eslint.tsconfig.json file was created