fix: remove "exists-in-wishlist" logic from search results
the wishlist store initially downloads only product counts. the actual products
are loaded later on demand, because that's a slow request.

the problem is if you search the wishlist while the actual wishlist products
aren't loaded yet, every search result will appear as if it's not in the
wishlist (heart icon instead of heart_fill)